backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
pydantic==2.6.4
pydantic-settings==2.2.1
httpx==0.27.0
python-dotenv==1.0.1
